api_yamdb/requirements.txt
requests==2.26.0
django==3.2
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
gunicorn==20.0.4
psycopg2-binary==2.9.5